@@ -1,7 +1,7 @@
 {
-  "name": "angular5_webpack",
+  "name": "angular6_webpack",
   "version": "1.0.0",
-  "description": "Angular 5 test using webpack",
+  "description": "Angular 6 demo using webpack",
   "scripts": {
     "build:dev": "rimraf wwwroot && webpack --config config/webpack.dev.js --progress --profile && copyfiles mock-authors.json favicon.ico wwwroot",
     "build:prod": "rimraf wwwroot && webpack --config config/webpack.prod.js --progress --profile --bail && copyfiles mock-authors.json favicon.ico wwwroot",
@@ -12,47 +12,47 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@angular/animations": "~5.2.7",
-    "@angular/cdk": "~5.2.3",
-    "@angular/common": "~5.2.7",
-    "@angular/compiler": "~5.2.7",
-    "@angular/core": "~5.2.7",
-    "@angular/forms": "~5.2.7",
-    "@angular/http": "~5.2.7",
-    "@angular/material": "~5.2.3",
-    "@angular/platform-browser": "~5.2.7",
-    "@angular/platform-browser-dynamic": "~5.2.7",
-    "@angular/router": "~5.2.7",
-    "angular-in-memory-web-api": "~0.5.3",
-    "core-js": "^2.5.3",
+    "@angular/animations": "~6.0.2",
+    "@angular/cdk": "~6.0.2",
+    "@angular/common": "~6.0.2",
+    "@angular/compiler": "~6.0.2",
+    "@angular/core": "~6.0.2",
+    "@angular/forms": "~6.0.2",
+    "@angular/http": "~6.0.2",
+    "@angular/material": "~6.0.2",
+    "@angular/platform-browser": "~6.0.2",
+    "@angular/platform-browser-dynamic": "~6.0.2",
+    "@angular/router": "~6.0.2",
+    "angular-in-memory-web-api": "~0.6.0",
+    "core-js": "^2.5.6",
     "hammerjs": "^2.0.8",
-    "rxjs": "5.5.6",
-    "zone.js": "^0.8.20"
+    "rxjs": "~6.1.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
     "@types/core-js": "^0.9.46",
     "@types/node": "^9.4.6",
     "angular2-template-loader": "^0.6.2",
     "awesome-typescript-loader": "^3.5.0",
-    "copyfiles": "^1.2.0",
-    "css-loader": "^0.28.10",
+    "copyfiles": "^2.0.0",
+    "css-loader": "^0.28.11",
     "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^1.1.11",
     "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^3.0.4",
+    "html-webpack-plugin": "^3.2.0",
     "null-loader": "^0.1.1",
     "open-browser-webpack-plugin": "0.0.5",
     "raw-loader": "^0.5.1",
     "rimraf": "^2.6.2",
-    "style-loader": "^0.20.2",
+    "style-loader": "^0.21.0",
     "to-string-loader": "^1.1.5",
-    "tslint": "^5.9.1",
-    "typescript": "^2.7.2",
+    "tslint": "^5.10.0",
+    "typescript": "^2.8.3",
     "webpack": "~3.10.0",
-    "webpack-cli": "~2.0.10",
+    "webpack-cli": "~2.1.3",
     "webpack-dev-server": "~2.11.1",
     "webpack-material-design-icons": "^0.1.0",
-    "webpack-merge": "^4.1.1"
+    "webpack-merge": "^4.1.2"
   },
   "repository": {}
 }
