@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 import { Author } from './author.model';
 
@@ -14,6 +15,7 @@ export class AuthorSearchService {
         return this.http
             //.get(`/mock-authors.json?name=${term}`)
             .get(`http://localhost:8081/api/authors/${term}`)
-            .map((response: any) => response.data as Author[]);
+            .pipe(
+            map((response: any) => response.data as Author[]));
     }
 }
